modules/grpc_api/requirements.txt
kafka-python
grpcio-tools
